+++ app/Livewire/Puntos/VentasNueva.php
@@ -1,7 +1,6 @@
 <?php
 
 namespace App\Livewire\Puntos;
-
 use Livewire\Component;
 
 class VentasNueva extends Component
