+++ app/Livewire/Recepcion/Ventas/Nueva/PagosTable.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire\Recepcion\Ventas\Nueva;
+
+use Livewire\Component;
+
+class PagosTable extends Component
+{
+    public function render()
+    {
+        return view('livewire.recepcion.ventas.nueva.pagos-table');
+    }
+}
