--- app/Livewire/Recepcion/VentasNueva.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Livewire\Recepcion;
-
-use Livewire\Component;
-
-class VentasNueva extends Component
-{
-    public function render()
-    {
-        return view('livewire.recepcion.ventas-nueva');
-    }
-}
